1_graph_creation/functions/agent5_writer/requirements.txt
functions-framework
flask
google-cloud-spanner
